Create data extractor
